fly_on_the_wall/pyproject.toml
[tool.poetry]
name = "fly_on_the_wall"
version = "0.1.0"
description = ""
authors = ["srini6773@gmail.com"]

[tool.poetry.dependencies]
python = "^3.8"
black = "^20.8b1"
boto = "^2.49.0"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
